Improve 3d dice dot drawing

+++ gtkprefs.c
@@ -1222,7 +1222,7 @@ void DoTestPerformance(GtkWidget *pw, GtkWidget* board)
 	BoardData *bd = BOARD(board)->board_data;
 	char str[255];
 	char *msg;
-	int fps;
+	float fps;
 	monitor m;
 
 	if (!GetInputYN(_("Save settings and test 3d performance for 3 seconds?")))
@@ -1248,7 +1248,7 @@ void DoTestPerformance(GtkWidget *pw, GtkWidget* board)
 	else
 		msg = _("3d Performance is very poor.\n");
 
-	sprintf(str, _("%s\n(%d frames per second)\n"), msg, fps);
+	sprintf(str, _("%s\n(%.1f frames per second)\n"), msg, fps);
 
 	outputl(str);
